+++ sdk/core/azure-core/inc/azure/core/http/curl_transport.hpp
@@ -40,6 +40,13 @@ namespace Azure { namespace Core { namespace Http {
   struct CurlTransportOptions final
   {
     /**
+     * @brief Default Maximum time in seconds that you allow the connection phase to the server to
+     * take.
+     *
+     */
+    AZ_CORE_DLLEXPORT static const long DefaultConnectionTimeout = 300;
+
+    /**
      * @brief The string for the proxy is passed directly to the libcurl handle without any parsing
      *
      * @remark No validation for the string is done by the Azure SDK. More about this option:
@@ -98,6 +105,15 @@ namespace Azure { namespace Core { namespace Http {
      *
      */
     bool NoSignal = false;
+
+    /**
+     * @brief Contain the maximum time in seconds that you allow the connection phase to the server
+     * to take.
+     *
+     * @details This only limits the connection phase, it has no impact once it has connected.
+     *
+     */
+    long ConnectionTimeout = DefaultConnectionTimeout;
   };
 
   /**